+++ controllers/posts/index.js
@@ -1,0 +1,9 @@
+const postBase = require("./postBase");
+const postLikes = require("./postLikes");
+const postComments = require("./postComments");
+
+module.exports = {
+  ...postBase,
+  ...postLikes,
+  ...postComments,
+};
